pyproject.toml
[tool.poetry]
name = "innotter"
version = "0.1.0"
description = ""
authors = ["sheirand <eugene.osakovich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
pip = "^22.1.2"
python-dotenv = "^0.20.0"
PyJWT = "^2.4.0"
pytest = "^7.1.2"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
